Implemented authentication checks on use cases.

+++ dux-manager-core/src/main/java/com/tracktainment/duxmanager/usecases/asset/DeleteAssetUseCase.java
@@ -1,6 +1,9 @@
 package com.tracktainment.duxmanager.usecases.asset;
 
 import com.tracktainment.duxmanager.dataprovider.AssetDataProvider;
+import com.tracktainment.duxmanager.exception.AuthenticationFailedException;
+import com.tracktainment.duxmanager.security.context.DigitalUserSecurityContext;
+import com.tracktainment.duxmanager.security.util.SecurityUtil;
 import lombok.*;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -11,8 +14,14 @@ import org.springframework.stereotype.Service;
 public class DeleteAssetUseCase {
 
     private final AssetDataProvider assetDataProvider;
+    private final SecurityUtil securityUtil;
 
     public void execute(Input input) {
+        // Check if digital user is authenticated
+        DigitalUserSecurityContext digitalUserSecurityContext = securityUtil.getDigitalUser();
+        if (!digitalUserSecurityContext.getId().equals(input.getDigitalUserId())) {
+            throw new AuthenticationFailedException("Authentication Failed: User ID does not match ID from JWT.");
+        }
         assetDataProvider.delete(
                 input.getDigitalUserId(),
                 input.getExternalId()
